reduce the size of arena keys

+++ crates/benchmarks/benches/benchmarks.rs
@@ -30,10 +30,10 @@ fn sounds(c: &mut Criterion) {
 	// a simple test case where many sounds are being played at once
 	c.bench_function("simple", |b| {
 		const SAMPLE_RATE: u32 = 48_000;
-		const NUM_SOUNDS: usize = 50_000;
+		const NUM_SOUNDS: u16 = 50_000;
 		let mut manager = AudioManager::<MockBackend>::new(AudioManagerSettings {
 			capacities: Capacities {
-				command_capacity: NUM_SOUNDS,
+				command_capacity: NUM_SOUNDS as usize,
 				sound_capacity: NUM_SOUNDS,
 				..Default::default()
 			},
@@ -56,10 +56,10 @@ fn sounds(c: &mut Criterion) {
 	// impact on the performance
 	c.bench_function("with on_start_processing callback", |b| {
 		const SAMPLE_RATE: u32 = 48_000;
-		const NUM_SOUNDS: usize = 50_000;
+		const NUM_SOUNDS: u16 = 50_000;
 		let mut manager = AudioManager::<MockBackend>::new(AudioManagerSettings {
 			capacities: Capacities {
-				command_capacity: NUM_SOUNDS,
+				command_capacity: NUM_SOUNDS as usize,
 				sound_capacity: NUM_SOUNDS,
 				..Default::default()
 			},
